+++ solstice/CHANGELOG.md
@@ -3,6 +3,10 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format.
 
 ## [Unreleased]
+### Added
+- Introduced `Capability` which takes OSGi `Provide-Capability`/`Require-Capability` into account. ([#71](https://github.com/equodev/equo-ide/pull/71))
+  - In particular, this means that we don't need to do any manual startup ordering anymore.
+  - Also reverted from SLF4J 2.x to 1.x because 2.x uses fancy parts of the capability system that nothing else in Eclipse seems to use.
 ### Fixed
 - No more errors on filesystems which don't support atomic move ([#73](https://github.com/equodev/equo-ide/pull/73))